paper add logging, set scaler initial value to 1

+++ pkg/haproxy/config.go
@@ -81,7 +81,7 @@ func createConfig(options options) *config {
 		brownout: &hatypes.Brownout{
 			ACLMapPath:        options.mapsDir + "/_brownout_rates.map",
 			Rates:             make(map[string]int),
-			UpdateDeployments: make(map[string]int),
+			UpdateDeployments: make(map[string]float64),
 		},
 		global:       &hatypes.Global{},
 		frontend:     &hatypes.Frontend{Name: "_front001"},
